DRY up fixtures across rule and tag tests
